Delete INSTALL
Information regarding external dependencies was not valid any more (because in this repository I added the version without external dependencies to the Microsoft Visual Studio Runtime DLLs, except for the system provided msvcrt.dll). The rest of the information is already provided in the README.md document - hence, delete this file.